Got a handle on our logging
The web service libraries we are using use java.util.logging.  This is a
bit of a pain to configure, so we use the jul-to-slf4j bridge so we can
handle it with logback.

For the tests we override with a logging.properties file, while for the
servlet we do it programmatically in the ServletContextClass.

Also added configuration to hide the policy warnings for now.  They have
been noted and are on the backlog.